restructures project and initiates css
