+++ src/Components/Login/Login.css
@@ -1,0 +1,1 @@
+.Login{position: absolute; top:50px}
