+++ src/Root/App.css
@@ -1,0 +1,4 @@
+.App {
+    bottom: 50px;
+}
+
